feat: Make neighbourhood links clickable
Clicking the links in the left panel has the same effect as clicking on
the marker.

Note, this causes a warning in React:

"Links must not point to "#". Use a more descriptive href or use a
button instead  jsx-a11y/href-no-hash"

This has been raised as an issue on GitHub, and it looks as though the
developers are considering removing this warning, so for the meantime,
this code will continue to throw the warning.  For more information,
see:

https://github.com/facebook/create-react-app/issues/4141