feat: Change focus of project
Dealing with crime data proved to be impossible because of the way the
pertinent data was fragmented across the API calls to the Police API.
This resulted in far too many requests and caused the app to be
painfully slow.  Instead of concentrating on crime data, the focus of
the app will be to find your local police station instead.