feat: Handle selected filtered neighbourhood
A selected neighbourhood will now be unselected if the filter results no
longer contain that neighbourhood.

For example, if the user selects "Wisbech" from the map, or
neighbourhood list, then types "q" in the filter, Wisbech will
automatically be deselected.  This creates a nice user experience as the
user will always be able to see the neighbourhood that they have
selected, plus the automatic panning also triggers in a more sensible
manner.