feat: Ensure smooth scrolling between markers
fitBounds() was called everytime the markers were filtered, which
caused a "hard" redraw of the map, causing it to flicker and jump to the
newly defined bounds.  panTo() has been implemented instead if there is
are filtered results, which causes smooth scrolling between the bounds.

The only time the focus "jumps" now, is when the number of filtered
results is the same as the number of total results.  This occurs on map
load, and also when the filter string is deleted by the user.